+++ api/src/main/java/org/mwdb/KType.java
@@ -8,26 +8,26 @@ public class KType {
     /**
      * Primitive Types
      */
-    public static final short BOOL = 1;
-    public static final short STRING = 2;
-    public static final short LONG = 3;
-    public static final short INT = 4;
-    public static final short DOUBLE = 5;
+    public static final byte BOOL = 1;
+    public static final byte STRING = 2;
+    public static final byte LONG = 3;
+    public static final byte INT = 4;
+    public static final byte DOUBLE = 5;
 
     /**
      * Primitive Arrays
      * SHOULD NOT BE USED OUTSIDE THE CORE IMPLEMENTATION.
      */
-    public static final short DOUBLE_ARRAY = 6;
-    public static final short LONG_ARRAY = 7;
-    public static final short INT_ARRAY = 8;
+    public static final byte DOUBLE_ARRAY = 6;
+    public static final byte LONG_ARRAY = 7;
+    public static final byte INT_ARRAY = 8;
 
     /**
      * Primitive Maps
      * SHOULD NOT BE USED OUTSIDE THE CORE IMPLEMENTATION.
      */
-    public static final short LONG_LONG_MAP = 9;
-    public static final short LONG_LONG_ARRAY_MAP = 10;
-    public static final short STRING_LONG_MAP = 11;
+    public static final byte LONG_LONG_MAP = 9;
+    public static final byte LONG_LONG_ARRAY_MAP = 10;
+    public static final byte STRING_LONG_MAP = 11;
 
 }
